Backend Authorization Configured

+++ src/server/database/queries/blogs.ts
@@ -6,9 +6,9 @@ import { IBlog, NewBlog, TagLink, ITag, sp } from '../../../types';
 
 const allBlogs = async () => Query<IBlog[]>('select * from blog');
 const oneBlog = async (id: number) => Query<IBlog[]>('select * from blog where id = ?', [id]);
-const editBlog = async (id: number, content: string) => Query('update blog set content = ? where id = ?', [content, id]);
-const deleteBlog = async (id: IBlog["id"]) => Query('delete from blog where id = ?', [id]);
-const addBlog = async (NewBlog: NewBlog) => Query('INSERT INTO blog set ?, authorid=1', [NewBlog]);
+const editBlog = async (id: number, content: string, userID: number) => Query('update blog set content = ? where id = ? AND userid = ?', [content, id, userID]);
+const deleteBlog = async (id: IBlog["id"], userID: number) => Query('delete from blog where id = ? AND userid= ?', [id, userID]);
+const addBlog = async (NewBlog: NewBlog, userID: number) => Query('INSERT INTO blog set ?, authorid=?', [NewBlog, userID]);
 // const addBlog = async (NewBlog: NewBlog, userID: number) => Query('INSERT INTO blog set ?, authorid=?', [NewBlog, userID]);
